fix review remark: remove `use super::*;`
